Do not lower the priority of DSO symbols than in-archive symbols
This commit reverts 0612ea41f1a4dad4439e898b1d2a675bfd1e85a2 and
instead solves the original issue with a different approach.

Now, if a hidden symbol was resolved to a DSO symbol, we'll redo
symbol resolution so that the symbol won't be resolved to the same one
next time.

This should fix our build-all buildbot.