New section-order feature: noinit
Allow the user to override the section types for custom input sections
This is rather common in micro controller projects:
to make data structures, that would otherwise be grouped to BSS,
magically appear in a special location, the structure is placed inside
a custom section which then gets defaulted to SHT_PROGBITS.
This then leads to an output section that would be all zeros in your final
binary. In gnu linker scripts this could be avoided by adding a NOLOAD
attribute to the section.

The use case above could be easily implemented in a driver - simply by setting
a pointer or writing a custom allocator.
Still every CPU vendor will give you overly complicated linker scripts examples.
This feature only makes sense for supporting older frameworks, and future MCU
vendor code.